chore(ssr): move transformers into dir
