+++ packages/@lwc/ssr-compiler/src/compile-template/transformers/component.ts
