+++ packages/@lwc/ssr-compiler/src/compile-template/transformers/for-of.ts
